Create Optimization Algorithm
The GUI prompts the user to input selected input parameters, such as room size, desired / current temperatures, # of occupants etc. The required cooling load and CADR are determined. The models dimensions / quantities are dynamically scaled optimizing size, weight and power requirements based on the required cooling load and filtration capacity. A text file is generated for reading by the Solidworks API, setting the value of each parameter, along with a log file for oversite during the algorithms execution.